Feature/getting movie by

+++ src/backend/MoviesRate.Domain/Entities/Review.cs
@@ -3,7 +3,7 @@
 public class Review : EntityBase
 {
     public Guid UserIdentifier { get; set; }
-    public long MovieId { get; set; }
+    public int MovieId { get; set; }
     public string Comments { get; set; } = string.Empty;
     public decimal Ratings { get; set; }
 }
